+++ lib/src/scanner.dart
@@ -1,7 +1,0 @@
-library scanner;
-
-import 'package:petitparser/petitparser.dart' hide Token;
-
-part 'parser_ext.dart';
-part 'token.dart';
-part 'tokenizer.dart';