+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next working
+
+- Tokens scanner & parser
+
 ## 0.0.2
 
 - Added simple tokenizer and tokens